+++ database/migrations/2025_05_04_072003_create_orders_table.php
@@ -11,13 +11,15 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('orders', function (Blueprint $table) {
-        $table->id();
-        $table->foreignId('user_id')->constrained()->onDelete('cascade');
-        $table->decimal('total', 10, 2);
-        $table->string('status')->default('pending'); // pending, paid, failed, shipped, etc
-        $table->timestamps();
-    });
+        if (!Schema::hasTable('orders')) {
+            Schema::create('orders', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('user_id')->constrained()->onDelete('cascade');
+                $table->decimal('total', 10, 2);
+                $table->string('status')->default('pending'); // pending, paid, failed, shipped, etc
+                $table->timestamps();
+            });
+        }
     }
 
     /**
