+++ database/migrations/2025_05_04_072025_create_order_items_table.php
@@ -11,15 +11,16 @@ return new class extends Migration
      */
     public function up(): void
     {
-       Schema::create('order_items', function (Blueprint $table) {
-        $table->id();
-        $table->foreignId('order_id')->constrained()->onDelete('cascade');
-        $table->foreignId('product_id')->constrained();
-        $table->integer('quantity');
-        $table->decimal('price', 10, 2); // preço no momento da compra
-        $table->timestamps();
-    });
-
+        if (!Schema::hasTable('order_items')) {
+            Schema::create('order_items', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('order_id')->constrained()->onDelete('cascade');
+                $table->foreignId('product_id')->constrained();
+                $table->integer('quantity');
+                $table->decimal('price', 10, 2); // preço no momento da compra
+                $table->timestamps();
+            });
+        }
     }
 
     /**
